Add Installer for Linux pyinstaller release version - add new dark iconbar icon
Remove Debugging from pyinstaller .spec files
Clean up make_linux
copy build_osx function to helpers/build_osx (not guaranteed to work)
add udev rules file for Linux installer

remove license/author information from readmes (too much repetition), rename license file
put 64bt before 32bit for (deprecated) py2app install
Note deprecation of setup-release.py

+++ contrib/encompass-release/helpers/build_osx.sh
@@ -1,11 +1,32 @@
+#!/bin/bash
+set -xeo pipefail
 VERSION="${1}"
-cd ..
-sed 's/ELECTRUM_VERSION/'${VERSION}'/g' Makefile.in > Makefile
-cd repo
-/opt/local/bin/python2.7 setup-release.py py2app
-mkdir ../src && mv dist/Encompass.app ../src/ 
-cd ..
-make
-mkdir -pv helpers/release-packages/OSX
-mv Encompass-${VERSION}.dmg helpers/release-packages/OSX
-mv src/Encompass.app helpers/release-packages/OSX
+test -z ${VERSION} && exit 1
+
+if [ "$(uname)" = "Darwin" ];
+   then
+  
+  if [ ! -f /opt/local/bin/python2.7 ]
+  then 
+    echo "This build requires macports python2.7 and pyqt4"
+    exit 5
+  fi  
+  VER="$1"
+  sed 's/ELECTRUM_VERSION/'${VER}'/g' Makefile.in > Makefile
+  cd repo
+  /opt/local/bin/python2.7 setup-release.py py2app
+  test -d ../src || mkdir ../src 
+  mv dist/Encompass.app ../src/ 
+  cd ..
+  #make  -  makes the unneeded dmg
+  test -d helpers/release-packages/OSX || mkdir -pv helpers/release-packages/OSX
+  #mv Encompass-${VER}.dmg helpers/release-packages/OSX
+  mv src/Encompass.app helpers/release-packages/OSX
+  cp helpers/make_OSX-installer.sh helpers/release-packages/OSX
+  thisdir=$(pwd)
+  cd helpers/release-packages/OSX
+  ./make_OSX-installer.sh $VERSION
+  cd ${thisdir}
+ else
+  echo "OSX Build Requires OSX build host!"                                                                                        
+ fi
